Big Update
FetchWinner is back so bad reign of WinnerLogic that ran 32787 loops 3 times a second is gone. I edited the VRCBilliards prefab a bit so I could get it to call my function whenever a win would occur so i am adding that file here aswell, I marked what I added with a comment.